Reorganize test files to mirror source file structure
- Create tests/test_models.py to properly test meal_planner/models.py
- Move makes model tests from test_models_makes.py to test_models.py
- Move makes UI tests from test_ui_makes.py to test_ui/test_recipe_editor.py
- Delete orphaned test_models_makes.py and test_ui_makes.py files
- Maintain 100% test coverage with all tests passing

Now test file structure properly mirrors source file structure.
